Introduce ValidatorBase
